feat(web): add display data and add choices to localstorage
